+++ src/cli.rs
@@ -39,7 +39,17 @@ pub fn build_cli() -> App<'static> {
                 .setting(clap::AppSettings::SubcommandRequiredElseHelp)
                 .display_order(0)
                 .about("Patient-level commands")
-                .subcommand(App::new("list").display_order(0).about("List all patients"))
+                .subcommand(
+                    App::new("list")
+                        .display_order(0)
+                        .about("List all patients")
+                        .arg(
+                            Arg::new("no_header")
+                            .about("Don't display table header")
+                            .short('n')
+                            .long("no-header"),
+                        )
+                )
                 .subcommand(
                     App::new("show")
                         .display_order(1)
@@ -60,7 +70,13 @@ pub fn build_cli() -> App<'static> {
                                 .about("Patient ID")
                                 .required(true)
                                 .value_name("ID"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                            .about("Don't display table header")
+                            .short('n')
+                            .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("search")
@@ -79,7 +95,13 @@ pub fn build_cli() -> App<'static> {
                                 .multiple_occurrences(true)
                                 .multiple_values(true)
                                 .value_name("QUERY"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("anonymize")
@@ -226,7 +248,17 @@ pub fn build_cli() -> App<'static> {
                 .setting(clap::AppSettings::SubcommandRequiredElseHelp)
                 .display_order(1)
                 .about("Study-level commands")
-                .subcommand(App::new("list").display_order(0).about("List all studies"))
+                .subcommand(
+                    App::new("list")
+                        .display_order(0)
+                        .about("List all studies")
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
+                    )
                 .subcommand(
                     App::new("show")
                         .display_order(1)
@@ -247,7 +279,13 @@ pub fn build_cli() -> App<'static> {
                                 .about("Study ID")
                                 .required(true)
                                 .value_name("ID"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("search")
@@ -266,7 +304,13 @@ pub fn build_cli() -> App<'static> {
                                 .multiple_occurrences(true)
                                 .multiple_values(true)
                                 .value_name("QUERY"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("anonymize")
@@ -408,7 +452,17 @@ pub fn build_cli() -> App<'static> {
                 .setting(clap::AppSettings::SubcommandRequiredElseHelp)
                 .display_order(2)
                 .about("Series-level commands")
-                .subcommand(App::new("list").display_order(0).about("List all series"))
+                .subcommand(
+                    App::new("list")
+                    .display_order(0)
+                    .about("List all series")
+                    .arg(
+                        Arg::new("no_header")
+                            .about("Don't display table header")
+                            .short('n')
+                            .long("no-header"),
+                    )
+                )
                 .subcommand(
                     App::new("show")
                         .display_order(1)
@@ -429,7 +483,13 @@ pub fn build_cli() -> App<'static> {
                                 .about("Series ID")
                                 .required(true)
                                 .value_name("ID"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("search")
@@ -448,7 +508,13 @@ pub fn build_cli() -> App<'static> {
                                 .multiple_occurrences(true)
                                 .multiple_values(true)
                                 .value_name("QUERY"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("anonymize")
@@ -598,7 +664,13 @@ pub fn build_cli() -> App<'static> {
                 .subcommand(
                     App::new("list")
                         .display_order(0)
-                        .about("List all instances"),
+                        .about("List all instances")
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("show")
@@ -639,7 +711,13 @@ pub fn build_cli() -> App<'static> {
                                 .multiple_occurrences(true)
                                 .multiple_values(true)
                                 .value_name("QUERY"),
-                        ),
+                        )
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("anonymize")
@@ -806,7 +884,13 @@ pub fn build_cli() -> App<'static> {
                 .subcommand(
                     App::new("list")
                         .display_order(0)
-                        .about("List all modalities"),
+                        .about("List all modalities")
+                        .arg(
+                            Arg::new("no_header")
+                                .about("Don't display table header")
+                                .short('n')
+                                .long("no-header"),
+                        )
                 )
                 .subcommand(
                     App::new("show")